+++ runa/widgets/radiobox_cfg.cpp
@@ -31,10 +31,10 @@ void runa::radiobox_cfg::init_widget(widget & _w, view_obj* _root_view) const
 void runa::radio_group_init(nana::radio_group& _group, std::vector<nana::checkbox*> _radios)
 {
     int selected = -1;
-    for (int i = 0; i < _radios.size(); ++i)
+    for (size_t i = 0; i < _radios.size(); ++i)
     {
         if (_radios[i]->checked())
-            selected = i;
+            selected = (int)i;
         _group.add(*_radios[i]);
     }
     if (selected != -1)
